[REFACTOR/#55] AlarmDay 도메인 레이어 이전 및 문자열 리소스 변환 로직 Feature 모듈로 이동
